refactor: SCRUM-117 footer 디렉토리 추가
