+++ src/components/footer/Footer.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Footer.css";
 
 const Footer = () => (
